+++ CHANGES.md
@@ -1,5 +1,9 @@
 # 20.08
 
+   * When using AMR and gravity or rotation, the source terms applied after
+     a reflux would have been incorrect if the previous timestep had a retry
+     (#1020). This has now been fixed. (#1112)
+
    * We now have the ability to access the problem-specific runtime
      parameters in C++ (#1093)